requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
alembic>=1.13
pydantic>=2.0
pydantic-settings>=2.0
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
openai>=1.40
python-dotenv>=1.0
python-multipart>=0.0.9
httpx>=0.27
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
aiofiles>=24.1
